requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
qrcode==7.4
Pillow
python-dateutil==2.8.2
